move bundle classes to util package
